Remove support for x-turbopuffer-flatten API
This was a hack to support the extra nesting in the `Product` operator.
We've simplified that operator, so we no longer need to support this.

(I don't want to keep it around in case we need it again in the future,
either, because the logic was not sufficiently general purpose, so
leaving it around is just inviting problems.)